fix(admin-ui): resolve circular dependency causing Vite warnings
