+++ packages/server-admin-ui/src/actions.ts
@@ -1,19 +1,8 @@
-import { isUndefined } from 'lodash'
 import { webSocketService } from './services/WebSocketService'
 import { useStore } from './store'
+import { authFetch, fetchAllData, fetchLoginStatus } from './dataFetching'
 
-declare global {
-  interface Window {
-    serverRoutesPrefix: string
-  }
-}
-
-const authFetch = (url: string, options?: RequestInit): Promise<Response> => {
-  return fetch(url, {
-    ...options,
-    credentials: 'include'
-  })
-}
+export { fetchAllData, fetchLoginStatus }
 
 export async function logoutAction(): Promise<void> {
   try {
@@ -157,48 +146,6 @@ export async function checkSecurityBackup(): Promise<boolean> {
     return data.hasBackup === true
   }
   return false
-}
-
-export async function fetchLoginStatus(): Promise<void> {
-  const response = await authFetch(`${window.serverRoutesPrefix}/loginStatus`)
-  if (response.status === 200) {
-    const data = await response.json()
-    useStore.getState().setLoginStatus(data)
-  }
-}
-
-export async function fetchAllData(): Promise<void> {
-  const fetchAndSet = async <T>(
-    endpoint: string,
-    setter: (data: T) => void,
-    prefix?: string
-  ) => {
-    try {
-      const response = await authFetch(
-        `${isUndefined(prefix) ? window.serverRoutesPrefix : prefix}${endpoint}`
-      )
-      if (response.status === 200) {
-        const data = await response.json()
-        setter(data)
-      }
-    } catch (error) {
-      console.error(`Failed to fetch ${endpoint}:`, error)
-    }
-  }
-
-  const state = useStore.getState()
-
-  await Promise.all([
-    fetchAndSet('/plugins', state.setPlugins),
-    fetchAndSet('/webapps', state.setWebapps),
-    fetchAndSet('/addons', state.setAddons),
-    fetchAndSet('/appstore/available', state.setAppStore),
-    fetchAndSet('/loginStatus', state.setLoginStatus),
-    fetchAndSet('/signalk', state.setServerSpecification, ''),
-    fetchAndSet('/security/access/requests', state.setAccessRequests),
-    fetchAndSet('/security/devices', state.setDevices),
-    fetchAndSet('/nodeInfo', state.setNodeInfo)
-  ])
 }
 
 export function openServerEventsConnection(isReconnect?: boolean): void {
