+++ packages/server-admin-ui/src/dataFetching.ts
@@ -1,0 +1,60 @@
+import { isUndefined } from 'lodash'
+import { useStore } from './store'
+
+declare global {
+  interface Window {
+    serverRoutesPrefix: string
+  }
+}
+
+export const authFetch = (
+  url: string,
+  options?: RequestInit
+): Promise<Response> => {
+  return fetch(url, {
+    ...options,
+    credentials: 'include'
+  })
+}
+
+export async function fetchLoginStatus(): Promise<void> {
+  const response = await authFetch(`${window.serverRoutesPrefix}/loginStatus`)
+  if (response.status === 200) {
+    const data = await response.json()
+    useStore.getState().setLoginStatus(data)
+  }
+}
+
+export async function fetchAllData(): Promise<void> {
+  const fetchAndSet = async <T>(
+    endpoint: string,
+    setter: (data: T) => void,
+    prefix?: string
+  ) => {
+    try {
+      const response = await authFetch(
+        `${isUndefined(prefix) ? window.serverRoutesPrefix : prefix}${endpoint}`
+      )
+      if (response.status === 200) {
+        const data = await response.json()
+        setter(data)
+      }
+    } catch (error) {
+      console.error(`Failed to fetch ${endpoint}:`, error)
+    }
+  }
+
+  const state = useStore.getState()
+
+  await Promise.all([
+    fetchAndSet('/plugins', state.setPlugins),
+    fetchAndSet('/webapps', state.setWebapps),
+    fetchAndSet('/addons', state.setAddons),
+    fetchAndSet('/appstore/available', state.setAppStore),
+    fetchAndSet('/loginStatus', state.setLoginStatus),
+    fetchAndSet('/signalk', state.setServerSpecification, ''),
+    fetchAndSet('/security/access/requests', state.setAccessRequests),
+    fetchAndSet('/security/devices', state.setDevices),
+    fetchAndSet('/nodeInfo', state.setNodeInfo)
+  ])
+}
